положил тесты в отдельный файл, добавил дефайны
