+++ lab2/dynamic_array.cpp
@@ -1,35 +1,9 @@
+#include "array_functions.h"
+#include "tests.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <assert.h>
 #include <time.h>
-
-
-typedef int elem_t;
-
-struct Stack
-{
-    elem_t* array;
-    size_t size;
-    size_t capacity;
-};
-
-struct Stack* stack_ctr(size_t size, size_t element_size);
-
-int push(struct Stack* st, void* buffer);
-
-int top(struct Stack* st, void* buffer);
-
-int pop(struct Stack* st);
-
-struct Stack* stack_dtr(struct Stack* st);
-
-void print_stack(struct Stack* st);
-
-void test1(struct Stack* st, int start);
-
-void test2(struct Stack* st, int start);
-
-void test3(struct Stack* st, int start);
 
 
 int main() {
@@ -65,7 +39,6 @@ struct Stack* stack_ctr(size_t size, size_t element_size) {
 
     if (!st) {
 
-        printf("No memory\n");
         return NULL;
     }
 
@@ -73,7 +46,6 @@ struct Stack* stack_ctr(size_t size, size_t element_size) {
 
     if (!st->array) {
 
-        printf("No memory\n");
         return NULL;
     }
 
@@ -98,9 +70,7 @@ int push(struct Stack* st, void* buffer) {
 
         if (!st->array) {
 
-            printf("No memory\n");
-
-            return 0;
+            return ERROR;
         }
 
         st->capacity *= COEF_EXPANSION;
@@ -108,7 +78,7 @@ int push(struct Stack* st, void* buffer) {
 
     st->array[st->size++] = *((elem_t*) buffer);
 
-    return 1;
+    return OK;
 }
 
 
@@ -120,12 +90,12 @@ int top(struct Stack* st, void* buffer) {
 
     if (st->size == 0) {
 
-        return 0;
+        return ERROR;
     }
 
     *((elem_t*) buffer) = st->array[st->size - 1];
 
-    return 1;
+    return OK;
 }
 
 
@@ -138,7 +108,7 @@ int pop(struct Stack* st) {
 
     if (st->size == 0) {
 
-        return 0;
+        return ERROR;
     }
 
     --st->size;
@@ -150,15 +120,13 @@ int pop(struct Stack* st) {
 
         if (!st->array) {
 
-            printf("No memory\n");
-
-            return 0;
+            return ERROR;
         }
 
         st->capacity = (size_t) (st->capacity / COEF_DECREASE);
     }
 
-    return 1;
+    return OK;
 }
 
 
@@ -187,88 +155,4 @@ void print_stack(struct Stack* st) {
     putchar('\n');
 
     printf("size = %u\ncapacity = %u\n", st->size, st->capacity);
-}
-
-
-void test1(struct Stack* st, int start) {
-
-    if (start) {
-    
-        int capacity = 1e6;
-
-        while (capacity < 1e5) {
-
-            for (int i = 0; i < (int) (capacity / 2); i++) {
-
-                pop(st);
-            }
-
-            for(int i = 0; i < (int) (capacity / 4); i++) {
-
-                push(st, &i);
-            }
-
-            capacity = (int) (capacity * 3 / 4);
-        }
-    }
-}
-
-
-void test2(struct Stack* st, int start) {
-
-    if (start) {
-
-        for (int i = 0; i < 100; i++) {
-
-            for (int j = 0; j < 10000; j++) {
-
-                pop(st);
-            }
-
-            for (int j = 0; j < 10000; j++) {
-
-                push(st, &i);
-            }
-        }
-
-        test1(st, 1);
-
-        for (int i = 0; i < 100; i++) {
-
-            for (int j = 0; j < 10000; j++) {
-
-                pop(st);
-            }
-
-            for (int j = 0; j < 10000; j++) {
-
-                push(st, &i);
-            }
-        }
-    }
-}
-
-
-void test3(struct Stack* st, int start) {
-
-    if (start) {
-
-        int operation = 0;
-        srand(start);
-
-        for (int i = 0; i < 1e6; i++) {
-            
-            operation = rand();
-
-            if (operation % 2 == 0) {
-
-                push(st, &i);
-            }
-
-            else {
-
-                pop(st);
-            }
-        }
-    }
 }
