+++ lab2/makefile_list
@@ -1,0 +1,19 @@
+CC = g++
+
+CFLAGS = -c -Wshadow -Winit-self -Wredundant-decls -Wcast-align -Wundef -Wfloat-equal -Winline -Wunreachable-code -Wmissing-declarations -Wmissing-include-dirs -Wswitch-enum -Wswitch-default -Weffc++ -Wmain -Wextra -Wall -g -pipe -fexceptions -Wcast-qual -Wconversion -Wctor-dtor-privacy -Wempty-body -Wformat-security -Wformat=2 -Wignored-qualifiers -Wlogical-op -Wno-missing-field-initializers -Wnon-virtual-dtor -Woverloaded-virtual -Wpointer-arith -Wsign-promo -Wstack-usage=8192 -Wstrict-aliasing -Wstrict-null-sentinel -Wtype-limits -Wwrite-strings -Werror=vla -D_DEBUG -D_EJUDGE_CLIENT_SIDE
+
+OBJECTS = single-linked_list.o tests.o
+
+all: single-linked_list.exe
+
+single-linked_list.exe: single-linked_list.o tests.o
+	$(CC) $(OBJECTS) -o single-linked_list.exe
+
+single-linked_list.o: single-linked_list.cpp array_functions.h
+	$(CC) $(CFLAGS) single-linked_list.cpp
+
+tests.o: tests.cpp tests.h
+	$(CC) $(CFLAGS) tests.cpp
+
+clean:
+	rm -rf *.o single-linked_list.exe
