+++ lab2/single-linked_list.cpp
@@ -1,39 +1,9 @@
+#include "list_functions.h"
+#include "tests.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <assert.h>
 #include <time.h>
-
-
-typedef int elem_t;
-
-struct Node
-{
-    elem_t data;
-    struct Node* prev;
-};
-
-struct Stack
-{
-    struct Node* head;
-};
-
-struct Stack* stack_ctr();
-
-int push(struct Stack* st, void* buffer);
-
-int top(struct Stack* st, void* buffer);
-
-int pop(struct Stack* st);
-
-struct Stack* stack_dtr(struct Stack* st);
-
-void print_stack(struct Stack* st);
-
-void test1(struct Stack* st, int start);
-
-void test2(struct Stack* st, int start);
-
-void test3(struct Stack* st, int start);
 
 
 int main() {
@@ -165,88 +135,4 @@ void print_stack(struct Stack* st) {
     }
 
     putchar('\n');
-}
-
-
-void test1(struct Stack* st, int start) {
-
-    if (start) {
-    
-        int capacity = 1e6;
-
-        while (capacity < 1e5) {
-
-            for (int i = 0; i < (int) (capacity / 2); i++) {
-
-                pop(st);
-            }
-
-            for(int i = 0; i < (int) (capacity / 4); i++) {
-
-                push(st, &i);
-            }
-
-            capacity = (int) (capacity * 3 / 4);
-        }
-    }
-}
-
-
-void test2(struct Stack* st, int start) {
-
-    if (start) {
-
-        for (int i = 0; i < 100; i++) {
-
-            for (int j = 0; j < 10000; j++) {
-
-                pop(st);
-            }
-
-            for (int j = 0; j < 10000; j++) {
-
-                push(st, &i);
-            }
-        }
-
-        test1(st, 1);
-
-        for (int i = 0; i < 100; i++) {
-
-            for (int j = 0; j < 10000; j++) {
-
-                pop(st);
-            }
-
-            for (int j = 0; j < 10000; j++) {
-
-                push(st, &i);
-            }
-        }
-    }
-}
-
-
-void test3(struct Stack* st, int start) {
-
-    if (start) {
-
-        int operation = 0;
-        srand(start);
-
-        for (int i = 0; i < 1e6; i++) {
-            
-            operation = rand();
-
-            if (operation % 2 == 0) {
-
-                push(st, &i);
-            }
-
-            else {
-
-                pop(st);
-            }
-        }
-    }
 }
