+++ lab2/tests.h
@@ -1,0 +1,13 @@
+#ifndef TESTS_H
+#define TESTS_H
+
+#define OK 1
+#define ERROR 0
+
+void test1(struct Stack* st, int start);
+
+void test2(struct Stack* st, int start);
+
+void test3(struct Stack* st, int start);
+
+#endif
